requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
comm==0.2.0
constantly==23.10.4
contourpy==1.1.0
cryptography==41.0.5
cssselect==1.2.0
cycler==0.11.0
dateparser==1.1.8
debugpy==1.8.0
decorator==5.1.1
demjson3==3.0.6
distro==1.8.0
duckdb==0.10.0
executing==2.0.1
extra-streamlit-components==0.1.56
facebook-scraper==0.2.59
fake-useragent==0.1.11
filelock==3.13.0
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.2.0
gcsfs==2024.2.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.27.1
ipython==8.18.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.8.1.78
packaging==23.1
pandas==1.5.3
parse==1.19.1
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.24.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.0
PyDispatcher==2.0.7
pyee==8.2.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==3.11.4
Pympler==1.0.1
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
queuelib==1.6.2
referencing==0.30.2
regex==2023.10.3
requests==2.26.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.6.2
seaborn==0.13.0
service-identity==23.1.0
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
sql_metadata==2.10.0
sqlparse==0.4.4
st-files-connection==0.1.0
st-gsheets-connection==0.0.4
stack-data==0.6.3
streamlit==1.26.0
streamlit-authenticator==0.2.2
tenacity==8.2.3
tesseract==0.1.3
threadpoolctl==3.2.0
tldextract==5.0.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
Twisted==23.8.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.18
validators==0.22.0
w3lib==2.1.2
wcwidth==0.2.12
websockets==10.4
weibo-scraper==1.0.6
xgboost==1.7.6
yarl==1.9.4
zipp==3.16.2
zope.interface==6.1
